requirements.txt
# fastapi
# uvicorn
# langchain
# langchain-ollama
# langchain-community
# langchain-google-genai
# langchain-huggingface
# google-generativeai
# openai
# faiss-cpu
# tiktoken
# python-dotenv
# sentence-transformers
# tf-keras

fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-google-genai
google-generativeai
sentence-transformers
faiss-cpu
langchain-ollama